+++ berserk/types/__init__.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from .account import AccountInformation, Perf, Preferences, Profile, StreamerInfo
 from .bulk_pairings import BulkPairing, BulkPairingGame
-from .common import ClockConfig, LightUser, OnlineLightUser
+from .common import ClockConfig, ExternalEngine, LightUser, OnlineLightUser
+from .puzzles import PuzzleRace
 from .opening_explorer import (
     OpeningExplorerRating,
     OpeningExplorerVariant,
@@ -10,14 +11,18 @@ from .opening_explorer import (
     Speed,
 )
 from .team import PaginatedTeams, Team
+from .tournaments import ArenaResult, CurrentTournaments, SwissResult, SwissInfo
 
 __all__ = [
     "AccountInformation",
+    "ArenaResult",
     "BulkPairing",
     "BulkPairingGame",
     "ClockConfig",
+    "ExternalEngine",
     "LightUser",
     "OnlineLightUser",
+    "CurrentTournaments",
     "OpeningExplorerRating",
     "OpeningExplorerVariant",
     "OpeningStatistic",
@@ -25,7 +30,10 @@ __all__ = [
     "Perf",
     "Preferences",
     "Profile",
+    "PuzzleRace",
     "Speed",
     "StreamerInfo",
+    "SwissInfo",
+    "SwissResult",
     "Team",
 ]