+++ CHANGELOG.rst
@@ -4,7 +4,27 @@ Changelog
 To be released
 --------------
 
-* Add `client.users.get_by_autocomplete`
+* Add ``client.explorer.get_otb_master_game``
+* Add::
+
+    client.external_engine.get
+    client.external_engine.get_by_id
+    client.external_engine.create
+    client.external_engine.update
+    client.external_engine.delete
+
+* Add::
+
+    client.tournaments.edit_swiss
+    client.tournaments.get_swiss
+    client.tournaments.join_swiss
+    client.tournaments.stream_swiss_results
+    client.tournaments.schedule_swiss_next_round
+    client.tournaments.terminate_swiss
+    client.tournaments.withdraw_swiss
+    
+* Add ``client.puzzles.create_race``
+* Add ``client.users.get_by_autocomplete``
 
 Thanks to @handsamtw and @Anupya for their contributions to this release.